traffic-generator/requirements.txt
confluent_kafka==1.9.0
kafka_python==2.0.2
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
protobuf==4.21.10
google-cloud-pubsub==2.13.11
PyYAML==6.0
